% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_firebase_timestamp.R
\name{convert_firebase_id}
\alias{convert_firebase_id}
\title{Convert a single firebase id to a timestamp}
\usage{
convert_firebase_id(id)
}
\arguments{
\item{id}{The firebase id (charachter).}
}
\value{
A date (POSIXct)
}
\description{
Convert a single firebase id to a timestamp
}
\references{
https://stackoverflow.com/questions/27908312/can-you-get-the-timestamp-from-a-firebase-array-key
}
\keyword{internal}
