% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_decode_timestamp.R
\name{convert_firebase_id_v}
\alias{convert_firebase_id_v}
\title{Vectorized version of \code{convert_firebase_id}}
\usage{
convert_firebase_id_v(id)
}
\description{
Vectorized version of \code{convert_firebase_id}
}
\keyword{internal}
