% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_firebase_timestamp.R
\name{decode_firebase_timestamp}
\alias{decode_firebase_timestamp}
\title{Decode a vector of firebase id into timestamp}
\usage{
decode_firebase_timestamp(id)
}
\arguments{
\item{id}{Firebase id (charachter).}
}
\value{
A date (POSIXct)
}
\description{
Decode a vector of firebase id into timestamp
}
