% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_decode_timestamp.R
\name{firebase_decode_timestamp}
\alias{firebase_decode_timestamp}
\title{Decode a vector of firebase id into timestamp}
\usage{
firebase_decode_timestamp(id)
}
\arguments{
\item{id}{Firebase id (charachter).}
}
\value{
A date (POSIXct)
}
\description{
Decode a vector of firebase id into timestamp
}
