% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebase_dowload.R
\name{firebase_download}
\alias{firebase_download}
\title{Download a firebase realtime data base.}
\usage{
firebase_download(url, ref = "/")
}
\arguments{
\item{url}{An URL pointing to a firebase realtime data base endpoint.}

\item{ref}{Defaults to "/".}
}
\value{
A list.
}
\description{
Download a firebase realtime data base.
}
