% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_to_pptx.R
\name{ggplot_to_pptx}
\alias{ggplot_to_pptx}
\title{Export a ggplot2 plot to a powerpoint document.}
\usage{
ggplot_to_pptx(plot = NULL, file, w_r = 1, template = NULL)
}
\arguments{
\item{plot}{Plot}

\item{file}{Path where the .pptx file should be saved.}

\item{w_r}{Wide to height ratio. Numeric.}

\item{template}{Powerpoint document template location. Defaults to an
internal powerpoint template.}
}
\description{
Export a ggplot2 plot to a powerpoint document.
}
