% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave.R
\name{ggsave}
\alias{ggsave}
\title{A \code{ggsave} alias}
\usage{
ggsave(
  filename,
  plot = ggplot2::last_plot(),
  device = NULL,
  path = NULL,
  scale = 1,
  width = 8.5,
  height = 4.75,
  units = "in",
  dpi = 600,
  limitsize = TRUE,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{device}{Device to use. Can either be a device function
(e.g. [png()]), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path of the directory to save plot to: `path` and `filename`
are combined to create the fully qualified file name. Defaults to the
working directory.}

\item{scale}{Multiplicative scaling factor.}

\item{width, height, units}{Plot size in `units` ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When `TRUE` (the default), `ggsave()` will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by `device`.}
}
\description{
\code{ggsave}, but with default values.
}
