% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_to_image.R
\name{latex_to_image}
\alias{latex_to_image}
\title{Download a latex equation as an image}
\usage{
latex_to_image(latex, path, format = "png")
}
\arguments{
\item{latex}{A latex equation (character).}

\item{path}{A path to save the image (character).}

\item{format}{The image format (supported: png, gif, svg).}
}
\value{
Invisibly returns the path.
}
\description{
Uses CodeCogs API to convert latex to an image, and then download the result.
}
