% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_es.R
\name{tidy_es}
\alias{tidy_es}
\title{Extension of broom::tidy with effect size and APA formatting.}
\usage{
tidy_es(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further argument to be passed to further methods.}
}
\value{
A tibble.
}
\description{
Extension of broom::tidy with effect size and APA formatting.
}
\details{
Currently, the \code{tidy_es} only supports \code{lm} objects.
}
